Create .kt
Explanation:

The var keyword is used to declare a mutable variable that can be reassigned to a different value.
In the first example, a var variable count is declared with the value 1. This variable can be reassigned to a different value, as shown in the next line.
In the second example, count is reassigned to the value 2, and then printed to the console.
In the third example, a var variable greeting is declared without an initial value. This variable can be assigned a value later, as shown in the next line.
In the fourth example, a var variable pi is declared with an initial value of 3.14. This variable can also be reassigned to a different value later.